Rewrite GUI with tkinter
Gooey is a great tool for quickly throwing together working and decent
looking GUIs, without having to know anything about GUI design or GUI
programming. Because of that it was a compelling choice for coub-gui.py,
since I never wrote a GUI before (and didn't plan to learn how to) but
still wanted to provide a basic GUI frontend for convenience.

While it was easy and fast to use, Gooey managed to become cumbersome in
its own way. I could nag about how the way CLI and GUI are designed is
fundamentally different and thus mapping CLI parsing to GUI interaction
will always include a considerable amount of hacks or how you can't
create checkboxes, which reliably show a check for True and no check for
False while allowing either as default value, but those are basically
nitpicks. My biggest gripe is all the trouble it caused me while working
on standalone binaries. On Linux it killed the idea entirely and even on
Windows it blew up the binaries considerably.

So for a while I entertained the notion to finally look into GUI
programming and replace the makeshift solution with a real one. This is
the result.

I decided to go with tkinter for several reasons.

  -) included in Python's standard library
  -) fairly lightweight
  -) native look on Windows and MacOS
  -) easy to package with PyInstaller
  -) easy to learn (not that I knew about this beforehand)

It has one big drawback though. It looks pretty dated (not to say ugly)
on Linux. Because of that PyGObject was another strong contender for me
(I really like the GTK look no matter the platform) and I may come back
to it in the future. But for now tkinter's advantages outweigh its dated
look on Linux.

As one could expect from a complete beginner, this initial version still
lacks a lot of polish in places. However it is in a usable state and
will allow me to shift some of my goals for this project by enabling me
to produce standalone builds for both Windows AND Linux (already
confirmed). Also unlike the previous GUI powered by Gooey, I consider
this iteration a central part of the project (otherwise I wouldn't have
invested so much time into it) and will eventually change the main code
base to better integrate the GUI and overall increase the quality of
this repo.