Overhaul input treeview
My initial goal was to simply add a scrollbar to the treeview widget, in
case more than 6 input sources (the fixed height of the widget) are
being used. But then one thing led to another and I overhauled much more
in one go than planned.

So let's break the changes down:

  -) fixed unwanted item deletion when the edit item window was closed

  -) split treeview's update_items() into add_item() and update_format()

Initially this method was used to add new items to the treeview and also
update its formatting (adjacent rows have different colors for a better
overview). Because the formatting must also be updated after the
deletion of items, update_items() was called by delete_item(), even
though the bulk of the method had nothing to do with it.

  -) move responsibility to handle treeview interactions to MainWindow

InputTree managed its own bindings and even called the edit item window
by itself. IMO the responsibility to handle these interactions should
lie with the window and not the widget itself. So the bindings were
moved to MainWindow's init and the edit item window calling logic is now
handled by a new method, which acts as callback for the edit item button
(just like for the new URL and new item button).

  -) scrollbar for InputTree

The initial goal of this commit. Shows up as soon as 7 or more items are
present in the treeview. Otherwise hidden.
Unfortunately this new scrollbar made me realize, that ScrolledText
still uses the old non-styled scrollbar and not the ttk version. So for
now we have two different scrollbar styles in this application. It also
means that I have to eventually replace ScrolledText with my own
implementation (which shouldn't be too difficult).

  -) new class to hold InputTree and its scrollbar

The MainWindow grid isn't suited to hold an optional scrollbar, so a new
frame class was introduced to handle the positioning of InputTree and
its scrollbar within the area which was previously inhibited only by
InputTree.

  -) More bindings

One of my goals for this GUI is being able to completely navigate it
with the keyboard. There's still a long road ahead of me, but adding,
editing and deleting items can now be done with keybinds (u -> new URL,
i -> new item, e -> edit selected item, delete key -> delete item).