Replace ScrolledText with own version
Like mentioned in the previous commit, ScrolledText (as provided from
the standard library) uses the old scrollbar instead of the new ttk
version.

Initially I planned to only replace the scrollbar type, but as it turned
out half the class is dedicated to a hack to infuse the frame's geometry
methods into the text widget. I don't see a point in repeating this.
So I decided to make ScrolledText a frame object, which holds a text and
scrollbar widget, instead of a text object, which holds a frame and a
scrollbar widget. Having to access the text widget for its methods
explicitly is also not more or less confusing compared to the original
ScrolledText.
The only aspect I kept is that any keyword arguments, used during the
construction, are passed to the text widget instead of the frame itself.

For now the scrollbars are still visible at all time (even if not
needed). Mainly because I don't know how to implement auto-hiding with
a textbox yet.