Untangle container formats from main script
Having all the code in a single file offers advantages for both users
and devs. The users don't have to worry about file structures and just
move around the script at will. The devs don't have to worry about the
availability of functions and variables (if global). Everything has
access to everything.

So putting everything in a script makes things easy. Easy is not
necessarily good though and it shouldn't excuse bad practices. Let's
look at the container classes for example. I think I described them as
self-contained at one point, but in reality they are deeply ingrained in
the overall execution. They handle their own terminal output, handle
their own retries and are even allowed to terminate the script. All of
that is only possible because they are part of a single file and have
access to the messaging system and the luxury of having the workflow
build around them.

But imagine one would want to pull out those containers and put them in
their own file. It breaks absolutely everything and could only be fixed
by tons of shared imports (between the new file and the main script) and
variable juggling.

So what's this wall of text supposed to convey? Basically that I am lazy
and built my script as the opposite of a modular program. Now that I
plan to split the script into more files (since I'm able to cover most
potential users with my standalone builds) it became apparent just how
bad some of my decisions were. It's time to remedy that and the
containers are the first part of the script to get this treatment.
Functions shared between argparse and their initialization are gone and
the retry logic and the messaging system are now part of parse_input.