Move container classes to own file
The first step to do away with the 1-file structure of this project.
2000 LoC are not that bad, but I can still organize the code base better
by splitting it into individual parts.

Because of the previous commit there was only little left to do to move
the classes to their own file. Previously readily available global
options are now passed on by other means. The biggest change in that
regard is that coubs_per_page isn't an advanced option anymore and
confined to container.py. There wasn't much of a reason to keep it. Too
high values will be ignored by Coub's API and behave like the max.
possible number. Too low values have no benefit and will only lead to
more required connection attempts.