Move default options class to own file
The next class to move. A lot less work was required to untangle it, so
it wasn't worth to split the two steps into separate commits.
The changes also contain a fix regarding an oversight in coub-gui.py as
I forgot to check for any config-related errors.

argparse-related classes and functions will likely move to the same
location in a later commit.