Move CLI-related code to options.py
As threatened in the last commit, all CLI-related code is now condensed
in a single file. This finally gets the main script below 1000 LoC,
although I'm still not done restructuring.

Looking at this commit, I must say that I totally forgot what a behemoth
this part of CoubDownloader is. So much code just to prevent the user
from shooting themselves in the foot.
I can't help but to ask myself, if it is really that much better than
the original custom parser? Well. Probably. The original parser was a
monster in its own right (oh god, all those custom validity checks).