Move processing logic for single coubs to own file
In contrast to the previous commits, this is quite messy. Mainly because
each Coub object is responsible for its own progress messages (easier to
do when processing them asynchronously) and there are tons of places,
where access to user options are required.
In a perfect world I would've done the same as with the containers and
overhauled the whole thing, before moving it to its own file. However,
we're not living in a perfect world, so my laziness got the better of
me (it was also encouraged by the fact that most pieces were in place to
easily allow me to move the whole thing without starting a domino effect
of missing resources, unlike with the containers). This means messy code
for now, which I will eventually refactor over time.