Combine GUI and CLI
GUI and CLI being separate entities had its reasons in the past. The CLI
was the main standalone script, while the GUI was a tacked-on feature.
But as more code was moved to their own files, coub.py quickly became
just another interface for the main bulk of the code.

So I just combined the two.

Calling the script without any arguments will launch the GUI.
Calling the script with arguments will use the CLI.
Calling the script with arguments and the -g/--gui option will launch
the GUI with the other settings applied (input from the command line is
ignore though).

Other than not being able to show the help text by calling the script
without arguments, there's only little impact on the CLI version. The
GUI, however, profits from it as it's now possible to call it with
command line options, thus giving the user a method to set temporary
options for the session instead of working with a config file.
There's also some code duplication I managed to get rid of and no need
for the makeshift GUI options object anymore.

The one drawback it has is that the resulting binary will be larger for
people, who only want the CLI version. However thanks to the switch from
Gooey to Tkinter, the difference between a CLI-only and a combined build
is only a few MB.

Whether the main window class will stay in the main script is doubtful,
but moving it will be a task for another day.