Combine process and attempt_process
Having two separate functions just to process a list of coubs (one to
gather the async tasks, one to handle retries via recursion) was rather
convoluted, not to mention that the recursion isn't really necessary.

So the two functions were combined to a single process function. Retries
are now handled by a simple while loop and the error messages and the
real exception handling are in one place.