Add option to output coub infos as JSON
Until now it was only possible to save additional information about a
coub by using a custom name template and including the wanted
information in the filename. The problem with this approach is that too
long filenames (easily achievable with %tags%) will fall back to %id%,
thus losing all additional data.

The option is very similar to youtube-dl's --print-json, only that it
writes to a file instead of stdout. No infos will be printed for
erroneous downloads, which also includes already existing and therefore
skipped files. Apart from that no additional duplicate checks are made
before outputting the infos to a file. I do not think it is within the
scope of this project to organize the additional data. We only provide
the information.

Fixes: #9