Remove GUI
This project isn't dead. Not that anybody had asked in the first place,
but it should still be said. Gyre is CoubDownloader's successor, but
sometimes having a CLI tool is just more convenient than a GUI. Plus
packaging a Python GTK program, that uses all kinds of GNOME-related
technologies, on Windows is still a pain.

Anyway, CoubDownloader has its place in the world, but not in its
current state. It will turn back into the simple CLI script it once was.
That means trimming a lot of the fat that accumulated over the years,
starting with the Tkinter GUI. One GUI already gives me enough headache,
so I don't need a second one on top of that. Maybe - but just maybe -
Gooey will be used in the future for another quick makeshift GUI, but
nothing more.