Overhaul settings system
The old way to make settings available everywhere was really clunky, not
to say outright bad. The options object was global to coub.py, but no
other module had access to it. So depending on the module, the main
script either changed global variables in those modules to reflect the
current settings or passed the object to every function/method that
needed it. It was a mess.

The new settings module takes heavy inspiration from Gyre's code. The
Settings class can now be imported if needed to get access to all
options. Simple as that. The approach clashed with argparse though, so
it was ripped out once again (and I don't miss it).

From here on out the code will be largely untested, broken and unfit for
usage, until I have done the majority of the overhaul.