Remove overwrite prompt
CoubDownloader is supposed to be a mass downloader for coubs. While an
overwrite prompt is useful for single coubs, it becomes a burden, when
you try to download a thousand and then the download gets stalled midway
because a single coub already existed. Or when you have to answer 500
prompts.

Instead the prompt setting gets converted to a simple overwrite rule.
Either the program is allowed to overwrite existing files or it will
skip them.