Change special treatment of "%id%" name template
Previously this special value was converted to None after CLI parsing.
Originally this was used to check for file existence before making any
connections to Coub, when the default name template was used (as ID is
the one special value that is known from the start). It made a huge
difference, when you had to skip over thousands of existing files while
using urllib. However, aiohttp is far faster when it comes to these
initial connections to read infos from the Coub API and if that is still
not fast enough then there's also the possibility of using an archive
file.
Meanwhile the fast existence check had its fair share of problems. Even
if we ignore the additional complexity to the overall processing
routine, it was often not 100% correct as the final file extension can
change based on stream availability (i.e. the original MP4 extension
will be kept, if no muxing is needed in the absence of audio) and that
is unknown in the beginning.

Now "%id%" alone is treated like any other name template